pyproject.toml
[project]
name = "dapla-toolbelt-metadata"
version = "0.13.0"
description = "Dapla Toolbelt Metadata"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Statistics Norway", email = "metadata@ssb.no" }]
maintainers = [
  { name = "Division for Data Enablement (724)", email = "metadata@ssb.no" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Typing :: Typed",
]
dependencies = [
  "arrow >=1.3.0",
  "beautifulsoup4 >=4.12.3",
  "dapla-auth-client>=1.2.1; python_version > '3.10'",
  "gcsfs >=2023.1.0",
  "google-auth >=2.38.0",
  "lxml >=5.3.1",
  "pandas>=2.3.3",
  "pyarrow >=8.0.0",
  "pydantic >=2.5.2",
  "pyjwt >=2.8.0",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv >=1.0.1",
  "pytz>=2025.2",
  "requests >=2.31.0",
  "ruamel-yaml >=0.18.10",
  "ssb-datadoc-model ==8.0.0",
  "ssb-klass-python >=1.0.1",
  "typing-extensions >=4.12.2",
  "universal-pathlib >=0.3.0",
  "urllib3>=2.5.0",
]

[project.urls]
Changelog = "https://github.com/statisticsnorway/dapla-toolbelt-metadata/releases"
Documentation = "https://statisticsnorway.github.io/dapla-toolbelt-metadata"
Homepage = "https://github.com/statisticsnorway/dapla-toolbelt-metadata"
Repository = "https://github.com/statisticsnorway/dapla-toolbelt-metadata"

[dependency-groups]
dev = ["pre-commit>=3.8.0", "ruff>=0.5.6"]
docs = [
  "furo>=2024.7.18",
  "myst-parser>=0.16.1",
  "sphinx>=6.2.1",
  "sphinx-autobuild>=2021.3.14",
  "sphinx-autodoc-typehints>=1.24.0",
  "sphinx-click>=3.0.2",
]
test = [
  "coverage[toml]>=6.2",
  "faker>=26.1.0",
  "pygments>=2.18.0",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=3.0.0",
  "pytest-mock>=3.14.0",
  "requests-mock>=1.12.1",
  "testcontainers[generic]>=4.8.2",
  "typeguard>=2.13.3",
  "xdoctest[colors]>=0.15.10",
]
type_check = [
  "mypy>=0.950",
  "pandas-stubs",
  "pyarrow-stubs>=10.0.1.9",
  "pytest>=8.3.2",
  "types-Pygments",
  "types-beautifulsoup4>=4.12.0.20240511",
  "types-colorama",
  "types-docker>=7.1.0.20250503",
  "types-requests",
  "types-setuptools",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool]
[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests", "*/test"]

[tool.coverage.run]
branch = true
source = ["dapla_metadata"]
omit = [
  "tests/*",
  "__init__.py",
  "*/dapla_metadata/variable_definitions/_generated/*"
]
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 80
omit = ["*/dapla_metadata/variable_definitions/_generated/*"]

[tool.hatch.build.targets.wheel]
packages = ["src/dapla_metadata"]

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true
# Generated code is not high quality, just ignore the problems there
exclude = "_generated/vardef_client/.*"

[[tool.mypy.overrides]]
module = ["dapla_metadata.variable_definitions._generated.vardef_client.*"]
warn_unreachable = false
ignore_errors = true
disable_error_code = ["unreachable"]

[[tool.mypy.overrides]]
# Allow missing type hints in third-party libraries without type information.
module = [
  "dapla",
  "faker",
  "gcsfs",
  "httpx",
  "nox",
  "pyarrow",
  "pyarrow.parquet",
  "pytest_mock",
  "ruamel.*",
  "testcontainers.*",
  "typeguard",
]
ignore_missing_imports = true

# Disable specific error codes in the 'tests' package
# Also don't require type annotations
[[tool.mypy.overrides]]
module = ["tests.*"]
disable_error_code = [
  "var-annotated",
  "has-type",
  "no-any-return",
  "no-untyped-def"
]

# tombi: lint.rules.one-of-multiple-match.disabled = true
# tombi: lint.rules.deprecated.disabled = true
[tool.pytest.ini_options]
filterwarnings = [
  "ignore::dapla_metadata.datasets.InconsistentDatasetsWarning",  # Our own
  "ignore::dapla_metadata.datasets.ObligatoryDatasetWarning",  # Our own
  "ignore::dapla_metadata.datasets.ObligatoryVariableWarning",  # Our own
  "ignore:^The @wait_container_is_ready decorator:DeprecationWarning",  # Internal third-party
]
pythonpath = ["src"]

[tool.ruff]
force-exclude = true  # Apply excludes to pre-commit
show-fixes = true
src = ["src", "tests"]
target-version = "py310"  # Minimum Python version supported
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]
extend-exclude = [
  "__pycache__",
  "old",
  ".ipynb_checkpoints",
  "noxfile.py",
  "docs/conf.py",
  "_generated/"
]

# Ruff rules may be customized as desired: https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN202",  # Don't requiere return type annotation for private functions.
  "ANN401",  # Allow type annotation with type Any.
  "COM812",  # Suggested to ignore when using ruff format
  "D100",  # Supress undocumented-public-module. Only doc of public api required.
  "FBT001",  # Allow boolean positional arguments in a function.
  "FBT002",  # Allow boolean default positional arguments in a function.
  "FIX002",  # Don't fail on TODO as long as it's documented correctly.
  "E402",  # Supress module-import-not-at-top-of-file, needed in jupyter notebooks.
  "E501",  # Supress line-too-long warnings: trust black's judgement on this one.
  "PLR2004",  # Allow to compare with unnamed numerical constants.
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"  # You can also use "numpy".

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
  "classmethod",
  "validator",
  "root_validator",
  "pydantic.validator"
]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"**/tests/*" = [
  "ANN001",  # type annotations don't add value for test functions
  "ANN002",  # type annotations don't add value for test functions
  "ANN003",  # type annotations don't add value for test functions
  "ANN201",  # type annotations don't add value for test functions
  "ANN204",  # type annotations don't add value for test functions
  "ANN205",  # type annotations don't add value for test functions
  "ANN206",  # type annotations don't add value for test functions
  "D100",  # docstrings are overkill for test functions
  "D101",
  "D102",
  "D103",
  "S101",  # asserts are encouraged in pytest
]

[tool.uv]
required-version = ">=0.8.0"
default-groups = ["dev", "test", "type_check"]
